Abstractions for ligand based event scoring

+++ pandda_gemmi/event/event_scoring_inbuilt.py
@@ -412,7 +412,7 @@ def get_interpolated_values_c(
     vals_list = gemmi.interpolate_pos_array(
         grid,
         transformed_structure_array.astype(np.float32),
-        vals.astype(np.float32),
+        vals
     )
     print(f"Vals list: {vals_list}")
     print(f"Vals: {vals}")
